requirements.txt
Django>=1.8
django-ipware>=1.1.1


